Added collision with bullets
Ragdolls were added to the enemies, and the HitBoxComponent script is added to them during initialization. When a bullet hits a ragdoll, a signal is triggered and the Ragdoll is activated. Bullets have a BulletComponent, which is responsible for velocity and VFX. All the bullets are in the object pool